+++ Editor/HorizontalSelectorEditor.cs
@@ -1,11 +1,11 @@
 #if UNITY_EDITOR
 using UnityEditor;
 using UnityEngine;
-using Jeomseon.UI.Components;
-using static Jeomseon.Editor.GUI.EditorGUILayoutActions;
-using static Jeomseon.Editor.SerializedPropertyReflection;
+using Jeomseon.Unity.UI.Components;
+using static Jeomseon.Unity.EditorToolkit.Editor.GUI.EditorGUILayoutActions;
+using static Jeomseon.Unity.EditorToolkit.Editor.SerializedPropertyReflection;
 
-namespace Jeomseon.UI.Editor
+namespace Jeomseon.Unity.UI.Editor
 {
 [CustomEditor(typeof(HorizontalSelector))]
 internal sealed class HorizontalSelectorEditor : UnityEditor.Editor