+++ Runtime/Components/HorizontalSelector.cs
@@ -9,7 +9,7 @@ using UnityEngine.Serialization;
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("com.jeomseon.ui.horizontal.selector.editor")]
 #endif
 
-namespace Jeomseon.UI.Components
+namespace Jeomseon.Unity.UI.Components
 {
     [DisallowMultipleComponent]
     public sealed class HorizontalSelector : MonoBehaviour