+++ Runtime/Components/RangeAdjustmentSlider.cs
@@ -9,7 +9,7 @@ using UnityEngine.Serialization;
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("com.jeomseon.ui.rangeadjustmentslider.editor")]
 #endif
 
-namespace Jeomseon.UI.Components
+namespace Jeomseon.Unity.UI.Components
 {
     using static UIHelper;